+ Adds: the logic in ActivityLogCardViewModelSlice
 → Moves: the activity log card logic to ActivityLogCardViewModelSlice
 from MySiteViewModel
